requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Login==0.6.3
Flask-WTF==1.2.1
email-validator==2.2.0
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
pytest==8.3.2
pytest-flask==1.3.0
WTForms==3.1.2
